Se realiza mejora

+++ domain/usecase/src/main/java/net/dev/carlosdev/usecase/managetasks/ManageTasksUseCase.java
@@ -8,16 +8,12 @@ public class ManageTasksUseCase {
 
     private final TasksRepository repository;
 
-    public String execute() {
-        return repository.findAll().toString();
-    }
-
     public Tasks createTask(Tasks task) {
         return repository.save(task);
     }
 
     public Tasks updateTask(Long id, Tasks task) {
-        if (repository.findById(id) != null) {
+        if (repository.findById(id) == null) {
             throw new RuntimeException("Task not found");
         }else {
             return repository.save(task);
@@ -25,11 +21,21 @@ public class ManageTasksUseCase {
     }
 
     public void deleteTask(Long id) {
-        repository.delete(id);
+        if (repository.findById(id) == null) {
+            throw new RuntimeException("Task not found");
+        }else {
+            repository.delete(id);
+        }
+
     }
 
     public Tasks getTask(Long id) {
-        return repository.findById(id);
+        if (repository.findById(id) == null) {
+            throw new RuntimeException("Task not found");
+        }else {
+            return repository.findById(id);
+        }
+
     }
 
     public Iterable<Tasks> getTasks() {
